Chatbot Navbar refactor
- in preparation for mobile views like planner + profile page, I've decided that we should offload a lot of our logic into separate places not just for code debugging but to extend the mobile navbar to other places as necessary
